refactor(history): extract reusable history screen
